% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_class}
\alias{check_class}
\title{Check an object's class}
\usage{
check_class(x, req)
}
\arguments{
\item{x}{An object.}

\item{req}{A character that defines the required class.}
}
\value{
If \code{x} does not inherit \code{req}, the function returns an error; otherwise \code{x} is invisibly returned.
}
\description{
This function checks an object's class.
}
\examples{
\dontrun{
check_class(as.numeric(1:10), "numeric") # works
check_class(1:10, "character")           # error
}
}
\author{
Edward Lavender
}
