% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Check a file exists}
\usage{
check_file_exists(x, type = c("abort", "warn"))
}
\arguments{
\item{x}{A string that defines a file path.}

\item{type}{A character that defines whether or not to throw a warning or an error.}
}
\value{
If the file \code{x} does not exist, the function returns a warning or error; otherwise a logical vector that defines whether or not each file exists is returned.
}
\description{
This function checks that a file exists.
}
\examples{
\dontrun{
check_file_exists(letters[1], type = "warn")
check_file_exists(letters[1:3], type = "warn")
check_file_exists(letters[1:3], type = "abort")
}
}
\author{
Edward Lavender
}
