% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv-package.R
\docType{package}
\name{dv}
\alias{dv-package}
\alias{dv}
\title{\code{dv}: Development tools for R}
\description{
\code{dv} provides tools to support the development of RStudio projects and R Packages.
\subsection{Directory helpers}{
\itemize{
\item \code{\link[=file_path]{file_path()}} and \code{\link[=here_path]{here_path()}} construct file paths;
\item \code{\link[=here_]{here_()}} functions are \code{\link[here:here]{here::here()}} wrappers for standard directories;
\item \code{\link[=repair_path]{repair_path()}}  checks file/directory path validity;
}
}

\subsection{RStudio Project tools (\verb{use_*})}{
\itemize{
\item \code{\link[=use_template_proj]{use_template_proj()}} creates a template RStudio Project;
\item \code{\link[=use_template_gitignore]{use_template_gitignore()}} creates a template .gitignore file;
\item \code{\link[=use_template_readme]{use_template_readme()}} creates a template README file;
\item \code{\link[=use_template_script]{use_template_script()}} creates a template script;
\item \code{\link[=use_template_tree]{use_template_tree()}} creates a directory tree;
}
}

\subsection{R package checks (\verb{check_*})}{
\itemize{
\item \code{\link{check_class}} checks an object's class
\item \code{\link[=check_dir_exists]{check_dir_exists()}} checks a directory exists;
\item \code{\link[=check_file_exists]{check_file_exists()}} checks a file exists;
}
}
}
\seealso{
\url{https://github.com/edwardlavender/dv}
}
\author{
Edward Lavender
}
