% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{file_path}
\alias{file_path}
\alias{.path}
\alias{here_path}
\title{Directory helper: construct file paths}
\usage{
.path(..., f)

file_path(...)

here_path(...)
}
\arguments{
\item{...}{Character vectors (see \code{\link[=file.path]{file.path()}} and \code{\link[here:here]{here::here()}}).}
}
\value{
The functions return \code{file.path(...)} or \code{here::here(...)}. If the path up to but excluding the last separator (i.e., \code{dirname(file.path(...))} or \code{dirname(here::here(...))}) does not exist, that part of the path is created as a side effect.
}
\description{
These functions are wrappers for \code{\link[=file.path]{file.path()}} and \code{\link[here:here]{here::here()}} plus \code{\link[=dir.create]{dir.create()}}. The functions require a character vector. If the directory name for the path does not exist, it is created recursively. The full path is then returned.
}
\details{
\code{.path()} is an internal function. \code{\link[=file_path]{file_path()}} and \code{\link[=here_path]{here_path()}} are exported. Both of the latter functions are designed to be coupled to a \code{save} function (e.g., \code{\link[=saveRDS]{saveRDS()}}) because they both create a directory (if necessary) and return the path, which can then be passed to the \code{save} function.
}
\examples{
#### Example (1): Implementation when the dirname exists
# The function simply returns the path
file_path(tempdir(), "tmp.rds")

#### Example (2): Implementation when the dirname does not exist
# The function creates the dirname (with a message) and returns the full path
file_path(tempdir(), "tmp_1", "tmp_2", "tmp.rds")
list.dirs(tempdir(), full.names = FALSE, recursive = TRUE)
# Reimplementing file_path() does not recreate existing directories
file_path(tempdir(), "tmp_1", "tmp_2", "tmp.rds")

#### Example (3): Implementation within a `save` function
# The function is useful within a save function because it
# ... creates the necessary directory before the file is saved
saveRDS("x", file_path(tempdir(), "tmp", "data", "x.rds"))
list.files(tempdir(), recursive = TRUE)
}
\seealso{
\code{\link{here_}}
}
\author{
Edward Lavender
}
\keyword{internal}
