% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-package.R
\name{glue_items}
\alias{glue_items}
\title{Glue items into a sentence}
\usage{
glue_items(text, items)
}
\arguments{
\item{text}{A character string.}

\item{items}{A character vector.}
}
\value{
The function returns a character string.
}
\description{
Glue items into a sentence
}
\details{
This is a simple wrapper for \code{\link[glue:glue]{glue::glue()}} that appends a list of items to an introductory sentence.
}
\examples{
\dontrun{

glue_items("This is a list of item(s):", letters[1:5])
}
}
\author{
Edward Lavender
}
\keyword{internal}
