% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{here_}
\alias{here_}
\alias{here_data_raw}
\alias{here_data}
\alias{here_r}
\alias{here_dev}
\alias{here_fig}
\alias{here_doc}
\title{Directory helper: \code{\link[here:here]{here::here()}} wrappers}
\usage{
here_data_raw(...)

here_data(...)

here_r(...)

here_dev(...)

here_fig(...)

here_doc(...)
}
\arguments{
\item{...}{Character vectors (see \code{\link[here:here]{here::here()}}).}
}
\value{
The functions return a file path.
}
\description{
These functions are simple wrappers for \code{\link[here:here]{here::here()}} for standard RStudio Project directories created by \code{\link[=use_template_proj]{use_template_proj()}}.
}
\examples{
here_data_raw()
here_data()
here_r()
here_dev()
here_fig()
here_doc()
}
\seealso{
\code{\link[=file_path]{file_path()}}, \code{\link[=here_path]{here_path()}}
}
\author{
Edward Lavender
}
