% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-projects.R
\name{open_sys}
\alias{open_sys}
\alias{.open_sys}
\title{Open a file}
\usage{
.open_sys(file)

open_sys(file)
}
\arguments{
\item{file}{A system path.}
}
\value{
The function is called for the side effect of opening a file.
}
\description{
This function opens input file(s) (or directories).
}
\author{
Edward Lavender
}
\keyword{internal}
