% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{repair_path}
\alias{repair_path}
\title{Directory helper: check path validity}
\usage{
repair_path(path)
}
\arguments{
\item{path}{A character string defining the file/directory path.}
}
\value{
The function returns a character string. If \code{path}, as inputted, is valid, it is returned unchanged. If \code{path} is invalid, the portion of the path that is valid is returned.
}
\description{
This function identifies whether or not a file or directory path is valid and, if not, determines the valid portion of the path.
}
\details{
This function is designed to facilitate quick identification of the valid/invalid components of manually specified paths (i.e., the point in the string at which the path becomes invalid).
}
\examples{
#### Example (1): Valid paths are returned unchanged
repair_path(tempdir())
file.create(file.path(tempdir(), "tmp.txt"))
repair_path(file.path(tempdir(), "tmp.txt"))

#### Example (2): If `path` is invalid, the valid portion of the path is returned
# Path broken at the 'top level'
repair_path("temp")
# Path broken after tempdir()
repair_path(file.path(tempdir(), "temp"))

#### Example (3): Examples with longer directories and files
# Generate new directory
dir.create(file.path(tempdir(), "temporary"), showWarnings = FALSE)
# Path broken after temporary/
repair_path(file.path(tempdir(), "temporary", "tmp2"))
# Create new file
file.create(file.path(tempdir(), "temporary", "file.txt"))
# File path is returned unchanged:
repair_path(file.path(tempdir(), "temporary", "file.txt"))
# Incorrectly specified path is broken after temporary/
repair_path(file.path(tempdir(), "temporary", "blah.txt"))

}
\author{
Edward Lavender
}
