% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-projects.R
\name{run}
\alias{run}
\title{Optionally run an expression}
\usage{
run(
  file,
  overwrite = FALSE,
  expr,
  read = readRDS,
  write = saveRDS,
  verbose = TRUE
)
}
\arguments{
\item{file}{A character that defines the file path.}

\item{overwrite}{A logical variable that defines whether or not to overwrite \code{file}, if it already exists. A \code{message} is given if so.}

\item{expr}{An expression to run.}

\item{read, write}{Read and write functions.}

\item{verbose}{A logical variable that defines whether or not to print the time.}
}
\value{
The function invisibly returns \code{NULL} or the output of \code{expr}.
}
\description{
This function optionally runs an expression & saves the output to a file, if the file does not exist (or \code{overwrite = TRUE}) or reads the file.
}
\details{
This function a simple wrapper for:
\verb{if (!file.exists(file)) \{ out <- eval(substitute(expr)) write(out) \} else \{ out <- read(file) \}}
}
\examples{
f <- tempfile()
a <- run(f, overwrite = FALSE, Sys.time())
Sys.sleep(0.5)
b <- run(f, overwrite = FALSE, Sys.time())
stopifnot(all.equal(a, b))

}
\author{
Edward Lavender
}
